Move tikxml bit
